requirements.txt
a2a==0.44
absl-py==2.3.1
aiodns==3.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.4
aiohttp-jinja2==1.6
aiosignal==1.3.2
ale-py==0.11.2
alembic==1.14.1
alpha_vantage==3.0.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
arrow==1.3.0
asyncio==3.4.3
attrs==25.3.0
audioop-lts==0.2.1
Automat==25.4.16
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
biiopython==0.0.5
black==25.1.0
blinker==1.9.0
boto3==1.39.1
botocore==1.39.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.2
coverage==7.9.2
cryptography==43.0.1
cssselect==1.3.0
curl_cffi==0.11.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
durationpy==0.10
faiss-cpu==1.11.0.post1
Farama-Notifications==0.0.4
fastapi==0.115.14
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
Flask==2.0.2
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.32.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-time==0.2.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-core==0.3.70
langchain-google-genai==2.1.8
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
litellm==1.74.7
loguru==0.7.2
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mccabe==0.7.0
mcp==1.12.1
mdurl==0.1.2
Mesa==3.2.0
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.11
mypy==1.16.1
mypy_extensions==1.1.0
neo4j==5.28.1
networkx==3.5
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.93.0
opencv-python==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pathspec==0.12.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
public==2020.12.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycares==4.9.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyflakes==3.4.0
pygame==2.6.1
Pygments==2.19.2
pyjabber==0.3.0
pymilvus==2.5.14
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.1
pytest-cov==6.2.1
python-a2a==0.5.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
queuelib==1.8.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.13.3
self==2020.12.3
sentence-transformers==5.0.0
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
Shimmy==2.0.0
singletonify==0.2.4
six==1.17.0
slixmpp==1.9.1
sniffio==1.3.1
soupsieve==2.7
spade==4.1.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
sse-starlette==2.4.1
stable_baselines3==2.6.0
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
swarms==8.0.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timeago==1.0.16
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.52.4
Twisted==25.5.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.4.0
uuid==1.30
uvicorn==0.35.0
w3lib==2.3.1
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
win32_setctime==1.2.0
winloop==0.1.8
yarl==1.20.1
yfinance==0.2.64
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
